[#141]: Implement group select

+++ engine/renderer/sprite.cpp
@@ -11,8 +11,8 @@ Sprite::~Sprite()
 {
    const auto transformMat = ComputeModelMat();
 
-   renderer::VulkanRenderer::SubmitMeshData(renderInfo_.idx, textures_, transformMat,
-                                            {0.0f, 0.0f, 0.0f, 0.0f});
+   renderer::VulkanRenderer::SubmitMeshData(static_cast< uint32_t >(renderInfo_.idx), textures_,
+                                            transformMat, {0.0f, 0.0f, 0.0f, 0.0f});
 
    renderer::VulkanRenderer::MeshDeleted(renderInfo_);
 }
@@ -26,7 +26,7 @@ Sprite::ComputeModelMat() const
 }
 
 void
-Sprite::ChangeRenderLayer(uint32_t newLayer)
+Sprite::ChangeRenderLayer(int32_t newLayer)
 {
    renderer::VulkanRenderer::MeshDeleted(renderInfo_);
 
@@ -130,8 +130,8 @@ Sprite::Render()
    {
       const auto transformMat = ComputeModelMat();
 
-      renderer::VulkanRenderer::SubmitMeshData(renderInfo_.idx, textures_, transformMat,
-                                               currentState_.color_);
+      renderer::VulkanRenderer::SubmitMeshData(static_cast< uint32_t >(renderInfo_.idx), textures_,
+                                               transformMat, currentState_.color_);
 
       changed_ = false;
    }
